Change fieldId type to long

+++ app/src/main/java/casak/ru/geofencer/domain/interactors/BuildArrowModelsInteractor.java
@@ -8,8 +8,8 @@ import casak.ru.geofencer.domain.model.ArrowModel;
 public interface BuildArrowModelsInteractor extends Interactor {
 
     interface Callback {
-        void onArrowsBuildFinished(int fieldId);
-        void onArrowsBuildFailed(int fieldId);
+        void onArrowsBuildFinished(long fieldId);
+        void onArrowsBuildFailed(long fieldId);
     }
 
     ArrowModel getLeftArrow();
